Fix PHP warning on admin-ajax request
Cf. https://sentry.greenpeace.org/organizations/greenpeace-org/issues/99/?project=2

For an admin-ajax request (background request done by the backend),
the global $menu is not yet defined on `admin_init`,
this leads to a warning raised on a WordPress function.

Switching the hook to `admin_menu` fixes the issue, and should keep the Options menu hidden.